Nettoyage des fichiers devenus obsolètes et schema image dédié dans un fichier
